spill out cpio

+++ src/RestoreRun.cpp
@@ -30,7 +30,8 @@
 using namespace std;
 
 RestoreRun::RestoreRun(RuntimeConfig& config, Repository& repository) :
-  repository(repository), config(config), numErrors(0), numFilesRestored(0), numBytesRestored(0)
+  repository(repository), config(config), numErrors(0), numFilesRestored(0), numBytesRestored(0),
+  fileCount(0)
 {
   // TODO Move out of constructor!!
   repository.lock();
@@ -52,21 +53,29 @@ void RestoreRun::restore(string& treeId, File& destinationDir, int depth)
       case TREEFILEENTRY_DIRECTORY: {
         File dir(destinationDir, entry.path);
 
-        bool skip = (config.skipExisting && dir.isDir());
-
-        if (!skip) {
-          if (config.verbose)
-            cout << "[d] " << dir.path << endl;
-
-          dir.mkdirs();
-        }
-        if (dir.isDir()) {
+        if (config.restoreAsCpio) {
+          printf("070707777777%06o%06o%06o%06o%06o%06o%011o%06o%011o%s%c",
+              ++fileCount, entry.fileMode, entry.uid, entry.gid, 1, 0,
+              (unsigned int) entry.mtime, (unsigned int) entry.path.size() +1, 0,
+              entry.path.substr(1).c_str(), 0x0);
           restore(entry.id, destinationDir, depth +1);
-          if (!skip) {
-            restoreMetaData(dir, entry);
-          }
         } else {
-          reportError(string("Cannot create destination directory: ").append(dir.path));
+          bool skip = (config.skipExisting && dir.isDir());
+
+          if (!skip) {
+            if (config.verbose)
+              cerr << "[d] " << dir.path << endl;
+
+            dir.mkdirs();
+          }
+          if (dir.isDir()) {
+            restore(entry.id, destinationDir, depth +1);
+            if (!skip) {
+              restoreMetaData(dir, entry);
+            }
+          } else {
+            reportError(string("Cannot create destination directory: ").append(dir.path));
+          }
         }
         break;
       }
@@ -74,28 +83,32 @@ void RestoreRun::restore(string& treeId, File& destinationDir, int depth)
       case TREEFILEENTRY_FILE: {
         File file(destinationDir, entry.path);
 
-        if (config.skipExisting && file.isFile()) break;
+        if (config.restoreAsCpio) {
 
-        if (config.verbose)
-          cout << "[f] " << file.path << endl;
+        } else {
+          if (config.skipExisting && file.isFile()) break;
 
-        // Create base directory:
-        if (depth == 0) {
-          file.getParent().mkdirs();
-          if (!file.getParent().isDir()) {
-            reportError(string("Cannot create destination directory: ").append(file.getParent().path));
+          if (config.verbose)
+            cerr << "[f] " << file.path << endl;
+
+          // Create base directory:
+          if (depth == 0) {
+            file.getParent().mkdirs();
+            if (!file.getParent().isDir()) {
+              reportError(string("Cannot create destination directory: ").append(file.getParent().path));
+            }
           }
-        }
-        file.remove();
+          file.remove();
 
-        try {
-          FileOutputStream out(file);
-          repository.exportFile(entry, out);
-          out.close();
-          restoreMetaData(file, entry);
-          numFilesRestored++;
-        } catch (Exception &ex) {
-          reportError(string("Cannot restore file ").append(file.path).append(": ").append(ex.getMessage()));
+          try {
+            FileOutputStream out(file);
+            repository.exportFile(entry, out);
+            out.close();
+            restoreMetaData(file, entry);
+            numFilesRestored++;
+          } catch (Exception &ex) {
+            reportError(string("Cannot restore file ").append(file.path).append(": ").append(ex.getMessage()));
+          }
         }
         break;
       }
@@ -103,23 +116,31 @@ void RestoreRun::restore(string& treeId, File& destinationDir, int depth)
       case TREEFILEENTRY_SYMLINK: {
         File file(destinationDir, entry.path);
 
-        if (config.skipExisting && file.isSymlink()) break;
+        if (config.restoreAsCpio) {
 
-        if (config.verbose)
-          cout << "[s] " << file.path << endl;
+        } else {
+          if (config.skipExisting && file.isSymlink()) break;
 
-        if (depth == 0)
-          file.getParent().mkdirs();
+          if (config.verbose)
+            cout << "[s] " << file.path << endl;
 
-        repository.exportSymlink(entry, file);
-        restoreMetaData(file, entry);
-        numFilesRestored++;
+          if (depth == 0)
+            file.getParent().mkdirs();
+
+          repository.exportSymlink(entry, file);
+          restoreMetaData(file, entry);
+          numFilesRestored++;
+        }
         break;
       }
 
       default:
         throw IllegalStateException("Unexpected tree file entry type");
     }
+  }
+
+  if (depth == 0 && config.restoreAsCpio) {
+    printf("0707070000000000000000000000000000000000010000000000000000000001300000000000TRAILER!!!");
   }
 }
